The prior commit did not fix the memory leak (apparently there was no real closure reference from the Results to the 'r' class/instance).
Jeff Rush indicated suspicion that one of the C extensions involved might not be releasing memory properly, but he has not had a chance to verify that. In the meantime I've created an experimental pure Python version of the Results class which does not rely on any C extensions such as Record.c or ExtensionClass. 

Unfortunately it is not passing one of the tests related to Acquisition, because I am not yet clear on how to implement that using pure Python. One Acquisition-related test has been commented out.

Three tests have been commented out related to string substitution, because the current implementation relies on namedtuple, which does not support dictionary style string substitution. I have yet to identify how to intercept string substitution so I can return the appropriate dictionary.

One test has been commented out because I did not see a reason for it. Why should slicing on a record raise an exception? The namedtuple can support slicing.

This has not been tested in a real application yet.

svn path=/Zope/branches/zeomega-2.11-memory-fixes/; revision=108778